Only release the GIL *after* calling Py_XXX
The previous version of the code released the GIL before calling
get_repr, but that's not OK because get_repr calls into PyObject_Repr,
which uses `Py_EnterRecursiveCall(" while getting the repr of an
object")`, which requires the to be held.

Moving the 'get_repr' before releasing the GIL fixes this.

Also, add "return duk_error", so it is clear that duk_error will never
return.

-------------------

This fixes the following segfault:

0  0x00007fffeee0ee00 in get_repr (value=value@entry=0x7fffef7707b8, buf=buf@entry=0x7fffeee5d0c0 <buf1> "", bufsz=bufsz@entry=200, value=<optimized out>, buf=<optimized out>, bufsz=<optimized out>) at src/conversions.c:5
1  0x00007fffeee3756f in python_function_caller (ctx=0x555555e5a4a0) at src/conversions.c:64
2  0x00007fffeee146ca in duk.handle_call_raw (thr=0x555555e5a4a0, idx_func=7, call_flags=8) at duk_js_call.c:2231